Phase Shifter script complete. Rewrote fundamental code structure for collision behavior
all colored platforms are now a part of ground layer but their tilemap collider component is disabled by default. Phase shifting enables the tilemap collider component on a linked platform thus allowing collision between the player and that platform. All colored platforms are in a list and their colliders get disabled when the player collides with a crystal that isn't linked to them.